feat: finish onboarding step on first upload or crawl
